Create creator class to manage the user's input to create persons or owners
